+++ test/src/forum/service/TopicService.java
@@ -12,15 +12,15 @@ public interface TopicService {
 	public TopicContent saveTopicContent(TopicContent topicContent);
 	
 	public void deleteTopic(Topic topic);
-	public void deleteTopicById(Integer topicId);
+	public void deleteTopicById(long topicId);
 	
 	public List<Topic> findTopicList(Topic s_topic, int pageSize,int pageNo);
 	
 	public Long getTopicCount(Topic s_topic);
 	
-	public Topic findTopicById(int topicId);	
+	public Topic findTopicById(long topicId);	
 	
-	public TopicContent findTopicContentByTopicId(int topicId);	
+	public TopicContent findTopicContentByTopicId(long topicId);	
 	
 	/**
 	 * 搜索置顶的帖子
@@ -29,7 +29,7 @@ public interface TopicService {
 	 * @param pageNo
 	 * @return
 	 */
-	public List<Topic> findZdTopicListBySectionId(int sectionId, int pageSize,int pageNo);
+	public List<Topic> findZdTopicListBySectionId(long sectionId, int pageSize,int pageNo);
 	
 	/**
 	 * 搜索精华的帖子
@@ -56,34 +56,34 @@ public interface TopicService {
 	 * @param pageNo
 	 * @return
 	 */
-	public List<Topic> findPtTopicListBySectionId(int sectionId, int pageSize,int pageNo);
+	public List<Topic> findPtTopicListBySectionId(long sectionId, int pageSize,int pageNo);
 	
-	public Long getPtTopicCountBySectionId(int sectionId);
+	public Long getPtTopicCountBySectionId(long sectionId);
 	
-	public List<Topic> findGoodTopicListBySectionId(int sectionId, int pageSize,int pageNo);
+	public List<Topic> findGoodTopicListBySectionId(long sectionId, int pageSize,int pageNo);
 	
-	public List<Topic> findNotGoodTopicListBySectionId(int sectionId, int pageSize,int pageNo);
+	public List<Topic> findNotGoodTopicListBySectionId(long sectionId, int pageSize,int pageNo);
 	
 	/**
 	 * 搜索未回复帖子
 	 * @param sectionId
 	 * @return
 	 */
-	public Long getNoReplyTopicCount(int sectionId);
+	public Long getNoReplyTopicCount(long sectionId);
 	
 	/**
 	 * 搜索精华帖子数
 	 * @param sectionId
 	 * @return
 	 */
-	public Long getGoodTopicCount(int sectionId);
+	public Long getGoodTopicCount(long sectionId);
 	
 	/**
 	 * 搜索帖子总数
 	 * @param sectionId
 	 * @return
 	 */
-	public Long getTotalTopicCount(int sectionId);
+	public Long getTotalTopicCount(long sectionId);
 	
 	/**
 	 * 搜索帖子总数
@@ -95,5 +95,5 @@ public interface TopicService {
 	 *获取帖子内容
 	 * @return
 	 */
-	public TopicContent getTopicContent(int topicId);
+	public TopicContent getTopicContent(long topicId);
 }
